requirements.txt
asgiref==3.8.1
django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
pyjwt==2.9.0
sqlparse==0.5.3
